Refactor policy processor to hide RemoveIf and introduce signal contexts and matchers
