+++ components/google-built-opentelemetry-collector/processor/googlepolicyprocessor/evaluator.go
@@ -23,9 +23,6 @@ import (
 	policyv1alpha1 "github.com/GoogleCloudPlatform/opentelemetry-operations-collector/gen/go/policy/v1alpha1"
 	"github.com/GoogleCloudPlatform/opentelemetry-operations-collector/pkg/googlepolicy"
 	"go.opentelemetry.io/collector/pdata/pcommon"
-	"go.opentelemetry.io/collector/pdata/plog"
-	"go.opentelemetry.io/collector/pdata/pmetric"
-	"go.opentelemetry.io/collector/pdata/ptrace"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -40,24 +37,6 @@ type Evaluator struct {
 	logPolicies    []*compiledLogPolicy
 	metricPolicies []*compiledMetricPolicy
 	tracePolicies  []*compiledTracePolicy
-}
-
-type compiledLogPolicy struct {
-	id       string
-	action   policyv1alpha1.Action
-	matchers []*compiledLogMatcher
-}
-
-type compiledMetricPolicy struct {
-	id       string
-	action   policyv1alpha1.Action
-	matchers []*compiledMetricMatcher
-}
-
-type compiledTracePolicy struct {
-	id       string
-	action   policyv1alpha1.Action
-	matchers []*compiledTraceMatcher
 }
 
 type predicateType int
@@ -83,21 +62,6 @@ type matcherPredicate struct {
 	lteVal      *policyv1alpha1.NumericValue
 	containsVal *policyv1alpha1.Value
 	negate      bool
-}
-
-type compiledLogMatcher struct {
-	target *policyv1alpha1.LogFieldSelector
-	pred   matcherPredicate
-}
-
-type compiledMetricMatcher struct {
-	target *policyv1alpha1.MetricFieldSelector
-	pred   matcherPredicate
-}
-
-type compiledTraceMatcher struct {
-	target *policyv1alpha1.TraceFieldSelector
-	pred   matcherPredicate
 }
 
 // NewEvaluator compiles a slice of TransformationPolicy objects into an Evaluator.
@@ -134,60 +98,6 @@ func NewEvaluator(policies []googlepolicy.TransformationPolicy) (*Evaluator, err
 		}
 	}
 	return ev, nil
-}
-
-func compileLogPolicy(p *policyv1alpha1.LogFilterPolicy) (*compiledLogPolicy, error) {
-	cp := &compiledLogPolicy{
-		id:     p.GetId(),
-		action: p.GetAction(),
-	}
-	for _, m := range p.GetMatches() {
-		pred, err := compilePredicate(m.GetPredicate(), m.GetNegate())
-		if err != nil {
-			return nil, err
-		}
-		cp.matchers = append(cp.matchers, &compiledLogMatcher{
-			target: m.GetTarget(),
-			pred:   pred,
-		})
-	}
-	return cp, nil
-}
-
-func compileMetricPolicy(p *policyv1alpha1.MetricFilterPolicy) (*compiledMetricPolicy, error) {
-	cp := &compiledMetricPolicy{
-		id:     p.GetId(),
-		action: p.GetAction(),
-	}
-	for _, m := range p.GetMatches() {
-		pred, err := compilePredicate(m.GetPredicate(), m.GetNegate())
-		if err != nil {
-			return nil, err
-		}
-		cp.matchers = append(cp.matchers, &compiledMetricMatcher{
-			target: m.GetTarget(),
-			pred:   pred,
-		})
-	}
-	return cp, nil
-}
-
-func compileTracePolicy(p *policyv1alpha1.TraceFilterPolicy) (*compiledTracePolicy, error) {
-	cp := &compiledTracePolicy{
-		id:     p.GetId(),
-		action: p.GetAction(),
-	}
-	for _, m := range p.GetMatches() {
-		pred, err := compilePredicate(m.GetPredicate(), m.GetNegate())
-		if err != nil {
-			return nil, err
-		}
-		cp.matchers = append(cp.matchers, &compiledTraceMatcher{
-			target: m.GetTarget(),
-			pred:   pred,
-		})
-	}
-	return cp, nil
 }
 
 func compilePredicate(predicate any, negate bool) (matcherPredicate, error) {
@@ -422,350 +332,6 @@ func numericToFloat64(nv *policyv1alpha1.NumericValue) float64 {
 		return 0
 	}
 }
-
-// ----------------------------------------------------------------------------
-// LOG EVALUATION
-// ----------------------------------------------------------------------------
-
-// EvalLog returns true if the log record should be DROPPED, false if it should be KEPT.
-func (e *Evaluator) EvalLog(lr plog.LogRecord, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string) bool {
-	if len(e.logPolicies) == 0 {
-		return false
-	}
-
-	var hasKeep, hasDrop bool
-	for _, p := range e.logPolicies {
-		if p.matches(lr, res, scope, resSchema, scopeSchema) {
-			if p.action == policyv1alpha1.Action_ACTION_KEEP {
-				hasKeep = true
-			} else if p.action == policyv1alpha1.Action_ACTION_DROP {
-				hasDrop = true
-			}
-		}
-	}
-
-	if hasKeep {
-		return false // KEEP always overrides DROP
-	}
-	if hasDrop {
-		return true // Drop matching record
-	}
-	return false // Default allow
-}
-
-func (p *compiledLogPolicy) matches(lr plog.LogRecord, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string) bool {
-	for _, m := range p.matchers {
-		val, exists := extractLogTarget(lr, res, scope, resSchema, scopeSchema, m.target)
-		if !m.pred.evaluate(val, exists) {
-			return false
-		}
-	}
-	return true
-}
-
-func extractLogTarget(lr plog.LogRecord, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string, target *policyv1alpha1.LogFieldSelector) (any, bool) {
-	if target == nil {
-		return nil, false
-	}
-	switch t := target.Target.(type) {
-	case *policyv1alpha1.LogFieldSelector_RecordField:
-		switch t.RecordField {
-		case policyv1alpha1.LogRecordField_LOG_RECORD_FIELD_BODY:
-			s := lr.Body().AsString()
-			return s, s != ""
-		case policyv1alpha1.LogRecordField_LOG_RECORD_FIELD_SEVERITY_TEXT:
-			s := lr.SeverityText()
-			return s, s != ""
-		case policyv1alpha1.LogRecordField_LOG_RECORD_FIELD_SEVERITY_NUMBER:
-			n := int64(lr.SeverityNumber())
-			return n, n != 0
-		case policyv1alpha1.LogRecordField_LOG_RECORD_FIELD_TRACE_ID:
-			tid := lr.TraceID()
-			if tid.IsEmpty() {
-				return nil, false
-			}
-			return tid.String(), true
-		case policyv1alpha1.LogRecordField_LOG_RECORD_FIELD_SPAN_ID:
-			sid := lr.SpanID()
-			if sid.IsEmpty() {
-				return nil, false
-			}
-			return sid.String(), true
-		default:
-			return nil, false
-		}
-	case *policyv1alpha1.LogFieldSelector_LogAttribute:
-		return lookupPath(lr.Attributes(), t.LogAttribute.GetPath())
-	case *policyv1alpha1.LogFieldSelector_ResourceAttribute:
-		return lookupPath(res.Attributes(), t.ResourceAttribute.GetPath())
-	case *policyv1alpha1.LogFieldSelector_ScopeAttribute:
-		return lookupPath(scope.Attributes(), t.ScopeAttribute.GetPath())
-	case *policyv1alpha1.LogFieldSelector_ScopeField:
-		switch t.ScopeField {
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_NAME:
-			s := scope.Name()
-			return s, s != ""
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_VERSION:
-			s := scope.Version()
-			return s, s != ""
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_SCHEMA_URL:
-			return scopeSchema, scopeSchema != ""
-		default:
-			return nil, false
-		}
-	default:
-		return nil, false
-	}
-}
-
-// ----------------------------------------------------------------------------
-// METRIC EVALUATION
-// ----------------------------------------------------------------------------
-
-// EvalMetricDatapoint returns true if the datapoint should be DROPPED, false if KEPT.
-func (e *Evaluator) EvalMetricDatapoint(m pmetric.Metric, dpAttrs pcommon.Map, temporality pmetric.AggregationTemporality, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string) bool {
-	if len(e.metricPolicies) == 0 {
-		return false
-	}
-
-	var hasKeep, hasDrop bool
-	for _, p := range e.metricPolicies {
-		if p.matches(m, dpAttrs, temporality, res, scope, resSchema, scopeSchema) {
-			if p.action == policyv1alpha1.Action_ACTION_KEEP {
-				hasKeep = true
-			} else if p.action == policyv1alpha1.Action_ACTION_DROP {
-				hasDrop = true
-			}
-		}
-	}
-
-	if hasKeep {
-		return false
-	}
-	if hasDrop {
-		return true
-	}
-	return false
-}
-
-func (p *compiledMetricPolicy) matches(m pmetric.Metric, dpAttrs pcommon.Map, temporality pmetric.AggregationTemporality, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string) bool {
-	for _, cm := range p.matchers {
-		val, exists := extractMetricTarget(m, dpAttrs, temporality, res, scope, resSchema, scopeSchema, cm.target)
-		if !cm.pred.evaluate(val, exists) {
-			return false
-		}
-	}
-	return true
-}
-
-func extractMetricTarget(m pmetric.Metric, dpAttrs pcommon.Map, temporality pmetric.AggregationTemporality, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string, target *policyv1alpha1.MetricFieldSelector) (any, bool) {
-	if target == nil {
-		return nil, false
-	}
-	switch t := target.Target.(type) {
-	case *policyv1alpha1.MetricFieldSelector_DescriptorField:
-		switch t.DescriptorField {
-		case policyv1alpha1.MetricDescriptorField_METRIC_DESCRIPTOR_FIELD_NAME:
-			s := m.Name()
-			return s, s != ""
-		case policyv1alpha1.MetricDescriptorField_METRIC_DESCRIPTOR_FIELD_DESCRIPTION:
-			s := m.Description()
-			return s, s != ""
-		case policyv1alpha1.MetricDescriptorField_METRIC_DESCRIPTOR_FIELD_UNIT:
-			s := m.Unit()
-			return s, s != ""
-		case policyv1alpha1.MetricDescriptorField_METRIC_DESCRIPTOR_FIELD_TYPE:
-			return metricTypeToString(m.Type()), true
-		case policyv1alpha1.MetricDescriptorField_METRIC_DESCRIPTOR_FIELD_AGGREGATION_TEMPORALITY:
-			return temporalityToString(temporality), true
-		default:
-			return nil, false
-		}
-	case *policyv1alpha1.MetricFieldSelector_DatapointAttribute:
-		return lookupPath(dpAttrs, t.DatapointAttribute.GetPath())
-	case *policyv1alpha1.MetricFieldSelector_ResourceAttribute:
-		return lookupPath(res.Attributes(), t.ResourceAttribute.GetPath())
-	case *policyv1alpha1.MetricFieldSelector_ScopeAttribute:
-		return lookupPath(scope.Attributes(), t.ScopeAttribute.GetPath())
-	case *policyv1alpha1.MetricFieldSelector_ScopeField:
-		switch t.ScopeField {
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_NAME:
-			s := scope.Name()
-			return s, s != ""
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_VERSION:
-			s := scope.Version()
-			return s, s != ""
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_SCHEMA_URL:
-			return scopeSchema, scopeSchema != ""
-		default:
-			return nil, false
-		}
-	default:
-		return nil, false
-	}
-}
-
-func metricTypeToString(t pmetric.MetricType) string {
-	switch t {
-	case pmetric.MetricTypeGauge:
-		return "gauge"
-	case pmetric.MetricTypeSum:
-		return "sum"
-	case pmetric.MetricTypeHistogram:
-		return "histogram"
-	case pmetric.MetricTypeExponentialHistogram:
-		return "exponential_histogram"
-	case pmetric.MetricTypeSummary:
-		return "summary"
-	default:
-		return "unspecified"
-	}
-}
-
-func temporalityToString(t pmetric.AggregationTemporality) string {
-	switch t {
-	case pmetric.AggregationTemporalityDelta:
-		return "delta"
-	case pmetric.AggregationTemporalityCumulative:
-		return "cumulative"
-	default:
-		return "unspecified"
-	}
-}
-
-// ----------------------------------------------------------------------------
-// TRACE EVALUATION
-// ----------------------------------------------------------------------------
-
-// EvalTraceSpan returns true if the span should be DROPPED, false if KEPT.
-func (e *Evaluator) EvalTraceSpan(span ptrace.Span, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string) bool {
-	if len(e.tracePolicies) == 0 {
-		return false
-	}
-
-	var hasKeep, hasDrop bool
-	for _, p := range e.tracePolicies {
-		if p.matches(span, res, scope, resSchema, scopeSchema) {
-			if p.action == policyv1alpha1.Action_ACTION_KEEP {
-				hasKeep = true
-			} else if p.action == policyv1alpha1.Action_ACTION_DROP {
-				hasDrop = true
-			}
-		}
-	}
-
-	if hasKeep {
-		return false
-	}
-	if hasDrop {
-		return true
-	}
-	return false
-}
-
-func (p *compiledTracePolicy) matches(span ptrace.Span, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string) bool {
-	for _, cm := range p.matchers {
-		val, exists := extractTraceTarget(span, res, scope, resSchema, scopeSchema, cm.target)
-		if !cm.pred.evaluate(val, exists) {
-			return false
-		}
-	}
-	return true
-}
-
-func extractTraceTarget(span ptrace.Span, res pcommon.Resource, scope pcommon.InstrumentationScope, resSchema, scopeSchema string, target *policyv1alpha1.TraceFieldSelector) (any, bool) {
-	if target == nil {
-		return nil, false
-	}
-	switch t := target.Target.(type) {
-	case *policyv1alpha1.TraceFieldSelector_RecordField:
-		switch t.RecordField {
-		case policyv1alpha1.SpanRecordField_SPAN_RECORD_FIELD_NAME:
-			s := span.Name()
-			return s, s != ""
-		case policyv1alpha1.SpanRecordField_SPAN_RECORD_FIELD_TRACE_ID:
-			tid := span.TraceID()
-			if tid.IsEmpty() {
-				return nil, false
-			}
-			return tid.String(), true
-		case policyv1alpha1.SpanRecordField_SPAN_RECORD_FIELD_SPAN_ID:
-			sid := span.SpanID()
-			if sid.IsEmpty() {
-				return nil, false
-			}
-			return sid.String(), true
-		case policyv1alpha1.SpanRecordField_SPAN_RECORD_FIELD_PARENT_SPAN_ID:
-			psid := span.ParentSpanID()
-			if psid.IsEmpty() {
-				return nil, false
-			}
-			return psid.String(), true
-		case policyv1alpha1.SpanRecordField_SPAN_RECORD_FIELD_STATUS_MESSAGE:
-			s := span.Status().Message()
-			return s, s != ""
-		case policyv1alpha1.SpanRecordField_SPAN_RECORD_FIELD_KIND:
-			return spanKindToString(span.Kind()), true
-		case policyv1alpha1.SpanRecordField_SPAN_RECORD_FIELD_STATUS_CODE:
-			return spanStatusToString(span.Status().Code()), true
-		default:
-			return nil, false
-		}
-	case *policyv1alpha1.TraceFieldSelector_SpanAttribute:
-		return lookupPath(span.Attributes(), t.SpanAttribute.GetPath())
-	case *policyv1alpha1.TraceFieldSelector_ResourceAttribute:
-		return lookupPath(res.Attributes(), t.ResourceAttribute.GetPath())
-	case *policyv1alpha1.TraceFieldSelector_ScopeAttribute:
-		return lookupPath(scope.Attributes(), t.ScopeAttribute.GetPath())
-	case *policyv1alpha1.TraceFieldSelector_ScopeField:
-		switch t.ScopeField {
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_NAME:
-			s := scope.Name()
-			return s, s != ""
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_VERSION:
-			s := scope.Version()
-			return s, s != ""
-		case policyv1alpha1.ScopeField_SCOPE_FIELD_SCHEMA_URL:
-			return scopeSchema, scopeSchema != ""
-		default:
-			return nil, false
-		}
-	default:
-		return nil, false
-	}
-}
-
-func spanKindToString(k ptrace.SpanKind) string {
-	switch k {
-	case ptrace.SpanKindInternal:
-		return "INTERNAL"
-	case ptrace.SpanKindServer:
-		return "SERVER"
-	case ptrace.SpanKindClient:
-		return "CLIENT"
-	case ptrace.SpanKindProducer:
-		return "PRODUCER"
-	case ptrace.SpanKindConsumer:
-		return "CONSUMER"
-	default:
-		return "INTERNAL"
-	}
-}
-
-func spanStatusToString(c ptrace.StatusCode) string {
-	switch c {
-	case ptrace.StatusCodeOk:
-		return "OK"
-	case ptrace.StatusCodeError:
-		return "ERROR"
-	default:
-		return "UNSET"
-	}
-}
-
-// ----------------------------------------------------------------------------
-// PATH LOOKUP HELPER
-// ----------------------------------------------------------------------------
 
 func lookupPath(attrs pcommon.Map, path []string) (any, bool) {
 	if len(path) == 0 {
